refactor(qwen3-reranker): move the context-aware collator into the model package
Every constant this collator carries is part of the Qwen3-Reranker prompt
contract -- the ChatML markers, the empty <think> block that makes the final
tokens a yes/no next-token prediction, the model card's system message, and
the yes/no label semantics. That is model-specific logic, which belongs under
components/models/<model>/ rather than the generic dataset collators.

Rename to Qwen3ContextAwareRerankerCollator and base it directly on
DataCollatorWithPadding. It overrode both methods CrossEncoderCollator
defines and used neither prompt_template nor format_text, so the old base was
vestigial; standing alone costs no duplication and drops the dependency on
the generic retrieval collator module.

The class is new on this branch and absent from origin/internal, so no
released _target_ path changes. Config references are updated in place.
